#314: Add reducer in app to store user infos
