+++ webapp/app/components/Login/index.js
@@ -5,7 +5,7 @@
 import React from 'react';
 import { PageHeader, Image } from 'react-bootstrap';
 import LoginForm from 'containers/Login/Form';
-import Logo from '../App/logo_sam_solution.png';
+import Logo from 'components/App/logo_sam_solution.png';
 import styles from './styles.css';
 
 /* eslint-disable react/prefer-stateless-function */
