+++ webapp/app/containers/App/constants.js
@@ -1,0 +1,5 @@
+//
+// App constants
+//
+
+export const SET_USER_INFO = 'SET_USER_INFO';
