+++ webapp/app/containers/Login/Form/actions.js
@@ -14,12 +14,9 @@ import request from 'utils/request';
 import { browserHistory } from 'react-router';
 import { resetStateUsername } from './Username/actions';
 import { resetStatePassword } from './Password/actions';
+import { setUserInfo } from 'containers/App/actions';
 
-
-import {
-  LOGIN,
-  SET_USER_INFO,
-} from './constants';
+import { LOGIN } from './constants';
 
 export function login(user) {
   return {
@@ -35,35 +32,8 @@ export function resetStateForm() {
   };
 }
 
-export function setUserInfo(logged, user) {
-  return {
-    type: SET_USER_INFO,
-    userInfo: {
-      logged,
-      userId: user.id,
-      username: user.name,
-      email: user.email,
-      isAdmin: user.isAdmin,
-    },
-  };
-}
-
-export function resetUserInfo() {
-  return {
-    type: SET_USER_INFO,
-    userInfo: {
-      logged: false,
-      userId: '',
-      username: '',
-      email: '',
-      isAdmin: '',
-    },
-  };
-}
-
 export function loginRequest(username, password) {
   return function returnLoginRequest(dispatch) {
-    console.log("TEST");
     return request
       .post('/api/public/user/login/')
       .type('form')
@@ -72,37 +42,7 @@ export function loginRequest(username, password) {
         if (!err && res.body.name) {
           dispatch(login(res.body));
           dispatch(setUserInfo(true, res.body));
-          console.log(`/edit-user/${res.body.id}`);
           browserHistory.push(`/edit-user/${res.body.id}`);
-        }
-      });
-  };
-}
-
-export function getUserInfo() {
-  return function startAction(dispatch) {
-    return request
-      .get('/api/logged-in/user/profile')
-      .end((err, res) => {
-        if (!err) {
-          dispatch(setUserInfo(true, res.body));
-        } else {
-          dispatch(resetUserInfo());
-        }
-      });
-  };
-}
-
-export function logoutRequest() {
-  return function startAction(dispatch) {
-    return request
-      .post('/api/logged-in/user/logout')
-      .end((err) => {
-        if (err) {
-          console.log(err);
-        } else {
-          dispatch(resetUserInfo());
-          browserHistory.push('/login');
         }
       });
   };
