+++ webapp/app/containers/RestoreHistory/index.js
@@ -8,7 +8,7 @@ import { RestoreHistory } from 'components/RestoreHistory';
 
 function mapStateToProps(state) {
   return {
-    userInfo: state.get('login').userInfo,
+    userInfo: state.get('app').userInfo,
   };
 }
