+++ webapp/app/containers/SaveCreation/Form/Users/index.js
@@ -10,7 +10,7 @@ import {
 
 function mapStateToProps(state) {
   return {
-    userInfo: state.get('login').userInfo,
+    userInfo: state.get('app').userInfo,
     users: state.get('saveCreation').get('SaveCreationFormUsersReducer').users,
     listAllUsers: state.get('save').get('SaveReducer').users,
     userError: state.get('saveCreation').get('SaveCreationFormUsersReducer').userError,
